requirements.txt
# ZnNi Report Generator - Web Application Dependencies

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Data Handling
pydantic>=2.0.0
pyyaml>=6.0

# PowerPoint Export (Phase 2)
python-pptx>=0.6.21
playwright==1.40.0

# Development & Testing
pytest>=7.4.3
httpx>=0.25.2